change value to not force float32 values

+++ Arm_test/arm_script.py
@@ -6,4 +6,4 @@ bus = SparkCAN.SparkBus(channel="can0", bustype='socketcan', bitrate=1000000)
 arm_motor = bus.init_controller(5)
 
 while (True):
-    arm_motor.percent_output(np.float32(0.05))
+    arm_motor.percent_output(0.05)
